Switch working directories on buffer change
This will put me in the same directory as a file, but then if it's in a
git repo will then move to the git repo root. Handy when jumping around
various directories/repos in one session so can run git operations or
commands from a location related to the current buffer.