Added Starship
This has a workaround to an issue with 0.48.0.

I've extracted the startup script with `/usr/local/bin/starship init bash --print-full-init`
and moved the CMD_STATUS earlier, then source it in the bash config.

I added this to the top to give an accurate shell start up time when
loading bash.