Fix Octo setup, gitgutter options for dotfiles, and command arguments
Need to run this setup when loading Octo.

Turns out autocmd patterns use `*` to descend subdirectories as well so this includes any folder in $HOME. Also missed the dot in .config, but didn't notice due to the wildcard in the other list item.

When creating a user command, you need to pass an argument list even if empty.