nvim: Fix file switcher
The `/` trigger pattern doesn't seem to work -- I thought it would match any file path because it causes file paths to expand to an absolute path, but maybe I misread the docs.

Let's instead just check we're in a directory and return early if not. Not yet sure if vim.fn.expand will succeed on non-file buffers yet, let's find out...